Added postal events profile.saved to the delete profile and create profile success handlers in the profile editor widget.
Note that the profile.saved is different from profile.changed (which does not exist yet). It implies
that the profile has been possibly changed and saved to storge, but the message does not carry
the actual profile data with it. The profile.changed message which also carries the changed profile
would be more efficient. Still, there is something to be said for the AppState getting the canonical
profile from storage for consumption by whatever widgets or components need it.