Add loggedin guard condition for the user profile page to prevent it from being seen from browser history.
After logout, although the kbase app will set the path to the login page, the user might still navigate back to
the profile page, e.g. using the back button. The profile page handles this -- not revealing any information from the
previous state and not attempting to get profile information for the nonexistent user -- but even better is to just
avoid the page. This change to the People controller sets the path to the login page if the session is not
authenticated.