Update Narrative Welcome Cell content as per https://atlassian.kbase.us/browse/NAR-697
The content for the welcome cell is now stored in a plain text file in normal markdown
format. Well, one difference is that there is an underscore template variable placeholder
for inserting the runtime doc site base url. This is available from the site config in
the javascript and is inserted into the content via underscore templates.
The markdown is converted to a snipped of javasript, which encodes the text as a concatenated
string (or rather a sequence of concatenations), which is encapsulated in an anonymous
function. The perl script (markdown-to-js-string.pl) is run with the text file as stdin,
and outputs the javscript, which can just be copied and pasted into the javscript file.
Also, the central config file needed updating to reflect the producton base url for the
doc site.