Minor css change in generic Heatmap vis widget
The class 'cell' is clashing with some CSS from iPython notebook and the narrative.  As far as I can tell, this class isn't actually being used anywhere by the heatmap widget, but so we can select it in the future I changed the name rather than removing the class altogether.

Tested locally in the Narrative, this seems to work.